moved comment test to WriteableTests
